Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
-float welcome() {
-    float a;
-    scanf("%f", &a);
+double welcome() {
+    double a;
+    scanf("%lf", &a);
     return a;
 }
 
 int main() {
-    printf("You entered: %f", welcome());
+    printf("You entered: %.4lf", welcome());
     return 0;
 }
